% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LENGDIR}
\alias{LENGDIR}
\title{The length classes by species}
\format{list}
\source{
\code{
attach('~/stasi/Splus5/SMB/.RData')
use_data(LENGDIR)
detach('file:~/stasi/Splus5/SMB/.RData')
}
}
\description{
A list whose names corresponds to species.id and whose element
are the length classes used to compute survey indices.
}
\author{
Hoskuldur Bjornsson
}

