% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lwc}
\alias{lwc}
\title{Length weigth coefficients}
\format{dataframe}
\source{
The
original source is found in "/net/hafkaldi/u2/reikn/Splus5/SMB/.RData". It
is copied into this library via:
\code{
names(LWCOEFF) <- iconv(names(LWCOEFF), "ISO8859-1", "UTF-8")
attributes(LWCOEFF$"12") <- NULL
lwc <- plyr::ldply(LWCOEFF)
names(lwc) <- c("species.id", "a","b")
lwc <-
  lwc %>%
  tbl_df() %>%
  mutate(species.id = as.integer(species.id)) %>%
  filter(!is.na(species.id))
use_data(lwc, overwrite = TRUE)
}
}
\description{
A \emph{dataframe} that contains length-weight coefficients
(a and b) for a number of species. Same data as \link{LWCOEFF}, the latter
being a list.
}
\author{
Hoskuldur Bjornsson
}

