% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LWCOEFF}
\alias{LWCOEFF}
\title{Length weigth coefficients}
\format{dataframe}
\source{
The
original source is found in "/net/hafkaldi/u2/reikn/Splus5/SMB/.RData". It
is copied into this library via:
\code{
attach("/net/hafkaldi/u2/reikn/Splus5/SMB/.RData")
names(LWCOEFF) <- iconv(names(LWCOEFF), "ISO8859-1", "UTF-8")
attributes(LWCOEFF$"12") <- NULL
use_data(LWCOEFF)
}
}
\description{
A \emph{list} that contains length-weight coefficients
(a and b) for a number of species. Same data as \link{lwc}, the latter
being a dataframe.
}
\author{
Hoskuldur Bjornsson
}

