% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hafvog.R
\name{read_hafvog}
\alias{read_hafvog}
\title{Reads data from Hafvog}
\usage{
read_hafvog(xmlfile)
}
\arguments{
\item{xmlfile}{Character vector specifying the name(s), including path
(location) of the xml-files.}
}
\value{
A list with the following dataframes:
\itemize{
\item st: Station table
\item le: Length frequency table
\item ot: Individual measurement table
\item ss: Subsampling (numer) table
}
}
\description{
Reads in xml-files generated by Hafvog, an electronic measurement
system used at Marine Research Institute, Reykavík.
}

