% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hafvog.R
\name{read_hafvog_counts}
\alias{read_hafvog_counts}
\title{Reads subsampling data from Hafvog}
\usage{
read_hafvog_counts(xmlfile)
}
\arguments{
\item{xmlfile}{Character vector specifying the name, including path
(location) of the xml-files.}
}
\value{
dataframe
}
\description{
Reads in subsamplin data from xml-files generated by
Hafvog, an electronic measurement system used at Marine Research Institute,
Reykavík.
}

