% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fjolst2.R
\name{read_lengths}
\alias{read_lengths}
\title{Title}
\usage{
read_lengths(id, species, col.names = c("id", "length", "n"), sex = NULL,
  mat = NULL, length = NULL, corrected = F, oracle = TRUE)
}
\arguments{
\item{id}{station id}

\item{species}{species id}

\item{col.names}{column names to return}

\item{sex}{include sex}

\item{mat}{include maturity}

\item{length}{specify length}

\item{corrected}{xx}

\item{oracle}{xx}
}
\value{
A dataframe
}
\description{
Title
}

