% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fjolst2.R
\name{read_stations}
\alias{read_stations}
\title{Title}
\usage{
read_stations(year = NULL, month = NULL, cruise.id = NULL,
  gear.id = NULL, rect = NULL, id = NULL, col.names, oracle = TRUE)
}
\arguments{
\item{oracle}{}
}
\description{
Title
}

