% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fjolst2.R
\name{read_subsampling}
\alias{read_subsampling}
\title{Title}
\usage{
read_subsampling(id, species, weight = FALSE, corrected = FALSE,
  oracle = TRUE)
}
\arguments{
\item{id}{xxx}

\item{species}{xxx}

\item{weight}{xxx}

\item{corrected}{xxx}

\item{oracle}{xxx}
}
\value{
A dataframe
}
\description{
Title
}

