% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smbSTODVAR}
\alias{smbSTODVAR}
\title{Synis.id and strata of the spring groundfish survey (SMB)}
\format{dataframe}
\description{
Contains only the strata for each tow, both the new and the
old stratas. The old strata is obtained from the column newstrata in
STODVAR located at
/net/hafkaldi/export/u2/reikn/Splus5/SMB/.RData. The newstrata is from
column newstrata in STODVAR located at /net/hafkaldi/export/u2/reikn/Splus5/SMBNewstrata/.RData
}

