% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smx_plots.R
\name{smx_4plot}
\alias{smx_4plot}
\title{smx_4plot}
\usage{
smx_4plot(tegund, dat, std = smx_standards)
}
\arguments{
\item{tegund}{character, like "torskur", "ysa", ....}

\item{dat}{data.frame}

\item{std}{data.frame containing some specs}
}

