% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratas_new}
\alias{stratas_new}
\title{Area of new strata}
\format{dataframe}
\source{
The object is created and maintained by Höskuldur Björnsson. The
original source is found in "/net/hafkaldi/export/u2/reikn/Splus5/HAUSTRALLNewStrata/Stratifiering/.RData". It
was copied into package via:
\code{
attach("/net/hafkaldi/export/u2/reikn/Splus5/HAUSTRALLNewStrata/Stratifiering/.RData")
x <- attributes(STRATAS)
stratas_new <-
  dplyr::data_frame(newstrata = 1:45,
                    area = as.numeric(x$area),
                    rall.area = x$rall.area,
                    lon = x$pos$lon,
                    lat = x$pos$lat,
                    .Names = x$names)
devtools::use_data(stratas_new)
}
}
\description{
Name, area size and central coordinate position of new strata
of the Icelandic groundfish survey.
}
\author{
Hoskuldur Bjornsson (hoski@hafro.is)
}

