% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratas_new_attributes}
\alias{stratas_new_attributes}
\title{Some title}
\format{Some format}
\source{
The object is created and maintained by Höskuldur Björnsson. The
original source is found in "/net/hafkaldi/u2/reikn/Splus5/SMB/.RData". It
is copied into this library via:
\code{
attach("/net/hafkaldi/export/u2/reikn/Splus5/HAUSTRALLNewStrata/Stratifiering/.RData")
x <- attributes(STRATAS)
stratas_new_attributes <-
  dplyr::data_frame(newstrata = 1:45,
                    area = as.numeric(x$area),
                    rall.area = x$rall.area,
                    lon = x$pos$lon,
                    lat = x$pos$lat,
                    .Names = x$names)
devtools::use_data(stratas_new_attributes)
}
}
\description{
from SMB RData
}
\author{
Hoskuldur Bjornsson (hoski@hafro.is)
}

