% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratas_old}
\alias{stratas_old}
\title{Area of old strata}
\format{dataframe}
\source{
The object is created and maintained by Höskuldur Björnsson. The
original source is found in "/net/hafkaldi/u2/reikn/Splus5/SMB/.RData". It
was copied into package via:
\code{
attach("~/stasi/Splus5/SMB/GEOMETRY.NEW/.RData")
x <- attributes(STRATAS)
stratas_old <- data_frame(oldstrata = as.integer(x$names),
                                 area = as.numeric(x$area),
                                 rall.area = x$rall.area,
                                 lon = x$pos$lon,
                                 lat = x$pos$lat,
                                 .Names = as.integer(x$.Names))
devtools::use_data(stratas_old)
}
}
\description{
Name, area size and central coordinate position of old strata
of the Icelandic groundfish survey.
}
\author{
Hoskuldur Bjornsson (hoski@hafro.is)
}

