% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratas_old_attributes}
\alias{stratas_old_attributes}
\title{Some title}
\format{Some format}
\source{
The object is created and maintained by Höskuldur Björnsson. The
original source is found in "/net/hafkaldi/u2/reikn/Splus5/SMB/.RData". It
is copied into this library via:
\code{
attach("/net/hafkaldi/u2/reikn/Splus5/SMB/GEOMETRY.NEW/.RData")
x <- attributes(STRATAS)
stratas_attributes <- data_frame(oldstrata = as.integer(x$names),
                                 area = as.numeric(x$area),
                                 rall.area = x$rall.area,
                                 lon = x$pos$lon,
                                 lat = x$pos$lat,
                                 .Names = as.integer(x$.Names))
devtools::use_data(stratas_attributes)
}
}
\description{
from SMB RData
}
\author{
Hoskuldur Bjornsson (hoski@hafro.is)
}

