% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datras.R
\name{tidy_length}
\alias{tidy_length}
\title{Title}
\usage{
tidy_length(hl, hh, correct_raising = FALSE)
}
\arguments{
\item{hl}{datras length frequency table}

\item{hh}{datras station table (cleaned)}

\item{correct_raising}{A boolean, if TRUE reconstruct original counts}
}
\value{
A dataframe
}
\description{
Title
}

