% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fjolst2.R
\name{read_lengths}
\alias{read_lengths}
\title{Title}
\usage{
read_lengths(id, species, col.names = c("id", "length", "n"), sex = NULL,
  mat = NULL, length = NULL, corrected = F, oracle = TRUE)
}
\arguments{
\item{oracle}{}
}
\description{
Title
}

