% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datras.R, R/datras2.R
\name{tidy_length}
\alias{tidy_length}
\title{Title}
\usage{
tidy_length(hl, hh, correct.raising = FALSE)

tidy_length(hl, hh, correct.raising = FALSE)
}
\arguments{
\item{hl}{datras length frequency table}

\item{hh}{datras station table (cleaned)}

\item{hl}{datras length frequency table}

\item{hh}{datras station table (cleaned)}
}
\description{
Title

Title
}

