% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datras.R, R/datras2.R
\name{tidy_station}
\alias{tidy_station}
\title{Tidy the datras station (HH) data}
\usage{
tidy_station(hh, to.lower = TRUE)

tidy_station(hh, to.lower = TRUE)
}
\arguments{
\item{hh}{datras station table}

\item{to.lower}{A boolean, if TRUE (default) then column names are all turned
to lower.}

\item{hh}{datras station table}

\item{to.lower}{A boolean, if TRUE (default) then column names are all turned
to lower.}
}
\description{
Creates a station id

Creates a station id
}

